feat: update example app to contain multiple screens
